requirements.txt
wheel
pyyaml
git+https://github.com/ammaraskar/pyCraft/